requirements.txt
aws-cdk-lib==2.108.0
constructs>=10.0.0,<11.0.0
streamlit>=1.28.0
botocore==1.31.85
boto3==1.28.85
aws-requests-auth==0.4.3